Split symbol preparation code out of symbol_bucket.
